+++ compiler/cli/src/org/jetbrains/kotlin/cli/jvm/K2JVMCompiler.kt
@@ -70,14 +70,8 @@ class K2JVMCompiler : CLICompiler<K2JVMCompilerArguments>() {
 
         setupCommonArgumentsAndServices(configuration, arguments, services)
 
-        try {
-            if (!arguments.noJdk) {
-                configuration.addJvmClasspathRoots(PathUtil.getJdkClassesRoots())
-            }
-        }
-        catch (t: Throwable) {
-            MessageCollectorUtil.reportException(messageCollector, t)
-            return INTERNAL_ERROR
+        setupJdkClasspathRoots(arguments, configuration, messageCollector).let {
+            if (it != OK) return it
         }
 
         try {
@@ -308,6 +302,41 @@ class K2JVMCompiler : CLICompiler<K2JVMCompilerArguments>() {
                 classpath.add(paths.runtimePath)
             }
             return classpath
+        }
+
+        private fun setupJdkClasspathRoots(arguments: K2JVMCompilerArguments, configuration: CompilerConfiguration, messageCollector: MessageCollector): ExitCode {
+            try {
+                if (!arguments.noJdk) {
+                    if (arguments.jdkHome != null) {
+                        messageCollector.report(CompilerMessageSeverity.LOGGING,
+                                                "Using JDK home directory ${arguments.jdkHome}",
+                                                CompilerMessageLocation.NO_LOCATION)
+                        val classesRoots = PathUtil.getJdkClassesRoots(File(arguments.jdkHome))
+                        if (classesRoots.isEmpty()) {
+                            messageCollector.report(CompilerMessageSeverity.ERROR,
+                                                    "No class roots are found in the JDK path: ${arguments.jdkHome}",
+                                                    CompilerMessageLocation.NO_LOCATION)
+                            return COMPILATION_ERROR
+                        }
+                        configuration.addJvmClasspathRoots(classesRoots)
+                    }
+                    else {
+                        configuration.addJvmClasspathRoots(PathUtil.getJdkClassesRoots())
+                    }
+                }
+                else {
+                    if (arguments.jdkHome != null) {
+                        messageCollector.report(CompilerMessageSeverity.WARNING,
+                                                "The '-jdk-home' option is ignored because '-no-jdk' is specified",
+                                                CompilerMessageLocation.NO_LOCATION)
+                    }
+                }
+            }
+            catch (t: Throwable) {
+                MessageCollectorUtil.reportException(messageCollector, t)
+                return INTERNAL_ERROR
+            }
+            return OK
         }
     }
 }